Added MeshModal component
