+++ src/components/index.ts
@@ -3,3 +3,4 @@ export { default as Button } from "./Button";
 export { default as Mesh } from "./Mesh";
 export { default as MeshMember } from "./MeshMember";
 export { default as Modal } from "./Modal";
+export { default as MeshModal } from "./MeshModal";
